TIP-842: Move group type repository to structure sub context
