+++ src/Akeneo/Pim/Structure/Bundle/Resources/config/repositories.yml
@@ -107,3 +107,17 @@ services:
         arguments:
             - '@doctrine.orm.entity_manager'
             - '%pim_catalog.entity.family_variant.class%'
+
+    pim_catalog.repository.group_type:
+        class: 'Akeneo\Pim\Structure\Bundle\Doctrine\ORM\Repository\GroupTypeRepository'
+        factory: 'doctrine.orm.entity_manager:getRepository'
+        arguments: ['%pim_catalog.entity.group_type.class%']
+        tags:
+            - { name: 'pim_repository' }
+
+    pim_enrich.repository.group_type:
+        class: 'Akeneo\Pim\Structure\Bundle\Doctrine\ORM\Repository\InternalApi\GroupTypeRepository'
+        arguments:
+            - '@pim_user.context.user'
+            - '@doctrine.orm.entity_manager'
+            - '%pim_catalog.entity.group_type.class%'
